+++ crates/web-client/js/__tests__/constants.test.js
@@ -37,6 +37,12 @@ describe("MethodName", () => {
     expect(MethodName.SUBMIT_NEW_TRANSACTION_MOCK).toBe(
       "submitNewTransactionMock"
     );
+    expect(MethodName.SUBMIT_NEW_TRANSACTION_BATCH).toBe(
+      "submitNewTransactionBatch"
+    );
+    expect(MethodName.SUBMIT_NEW_TRANSACTION_BATCH_MOCK).toBe(
+      "submitNewTransactionBatchMock"
+    );
     expect(MethodName.SUBMIT_NEW_TRANSACTION_WITH_PROVER).toBe(
       "submitNewTransactionWithProver"
     );